Seperate page for Modifying Terms, working save and delete

+++ WGUMobilePortal/ViewModels/TermsViewModel.cs
@@ -2,6 +2,7 @@
 using System.Collections.ObjectModel;
 
 using WGUMobilePortal.Services;
+using WGUMobilePortal.Views;
 
 using Xamarin.Forms;
 
@@ -10,9 +11,24 @@ namespace WGUMobilePortal.ViewModels
     public class TermsViewModel : BaseViewModel
     {
         public ObservableCollection<Models.Term> Term { get; set; }
+        //public Models.Term SelectedTerm
+        //{
+            //get => _selectedTerm;
+            //set
+            //{
+                //if (value != _selectedTerm)
+                //{
+                    //SetProperty(ref _selectedTerm, value);
+                    //OnPropertyChanged();
+                //}
+            //}
+        //}
         public Command RefreshCommand { get; }
         public Command AddCommand { get; }
         public Command<Models.Term> RemoveCommand { get; }
+        public Command ModifyCommand { get; }
+
+        //private Models.Term _selectedTerm;
 
 
         public TermsViewModel()
@@ -25,11 +41,13 @@ namespace WGUMobilePortal.ViewModels
             RefreshCommand = new Command(Refresh);
             AddCommand = new Command(Add);
             RemoveCommand = new Command<Models.Term>(Remove);
+            ModifyCommand = new Command<Models.Term>(Modify);
 
             Load();
         }
 
-        async void Add()
+
+            async void Add()
         {
             //string name = "Term 1";
             string name = await App.Current.MainPage.DisplayPromptAsync("Name", "Name of Term");
@@ -45,8 +63,11 @@ namespace WGUMobilePortal.ViewModels
 
         async void Remove(Models.Term term)
         {
-            await DBService.RemoveTerm(term.Id);
-            Refresh();
+            if (await Shell.Current.DisplayAlert("Confirm Deletion", $"Are you sure you want to delete {term.Name}", "Delete", "Cancel"))
+            {
+                await DBService.RemoveTerm(term.Id);
+                Refresh();
+            }
         }
 
         async void Refresh()
@@ -67,7 +88,7 @@ namespace WGUMobilePortal.ViewModels
             IsBusy = false;
         }
 
-        async void Load()
+        public async void Load()
         {
             IsBusy = true;
             Term.Clear();
@@ -77,6 +98,16 @@ namespace WGUMobilePortal.ViewModels
                 Term.Add(term);
             }
             IsBusy = false;
+        }
+
+        async void Modify(Models.Term term)
+        {
+            //SelectedTerm = term;
+            //var termsModifyPage = new Views.ModifyTermsPage();
+            //termsModifyPage.BindingContext = term;
+            await AppShell.Current.GoToAsync($"{nameof(ModifyTermsPage)}?id={term.Id}&name={term.Name}&startDate={term.StartDate}&endDate={term.EndDate}");
+                //($"///ModifyTermsPage"); //?Id={term.Id}");
+            //await AppShell.Current.Navigation.PushAsync(termsModifyPage);
         }
     }
 }
